requirements.prod.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
aiosqlite==0.21.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arq==0.26.1
arrow==1.4.0
asttokens==3.0.1
async-lru==2.2.0
asyncio-throttle==1.0.2
attrs==25.4.0
babel==2.18.0
backoff==2.2.1
beautifulsoup4==4.14.3
bleach==6.3.0
cachetools==4.2.1
catboost==1.2.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cryptography==44.0.2
cuda-bindings==12.9.4
cuda-pathfinder==1.4.3
cycler==0.12.1
datasets==4.8.2
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
dill==0.4.1
distro==1.9.0
drain3==0.9.11
executing==2.2.1
fastapi==0.115.12
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
google-auth==2.49.1
google-genai==1.66.0
graphviz==0.21
greenlet==3.3.2
grpcio==1.78.0
grpcio-tools==1.71.2
h11==0.16.0
h2==4.3.0
hf-xet==1.4.2
hiredis==3.3.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.7.1
hyperframe==6.1.0
idna==3.11
ipykernel==7.2.0
ipython==9.10.0
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
json5==0.13.0
jsonpatch==1.33
jsonpickle==1.5.1
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.8.0
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.4
jupyterlab==4.5.6
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
kiwisolver==1.5.0
langchain-core==1.2.20
langchain-google-genai==4.2.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-checkpoint-sqlite==3.0.3
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.20
lark==1.3.1
lightgbm==4.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mem0ai==1.0.5
missingno==0.5.2
mistune==3.2.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.18.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook_shim==0.2.4
numpy==2.4.3
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
openai==2.29.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.1
pandocfilters==1.5.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.4
plotly==6.6.0
portalocker==2.10.1
posthog==7.9.12
prometheus_client==0.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.6
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.8.1
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-json-logger==4.0.0
pytz==2026.1.post1
PyYAML==6.0.3
pyzmq==27.1.0
qdrant-client==1.13.3
rank-bm25==0.2.2
redis==5.2.1
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
Send2Trash==2.1.0
sentencepiece==0.2.1
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sqlite-vec==0.1.7
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
terminado==0.18.1
threadpoolctl==3.6.0
tinycss2==1.4.0
tokenizers==0.22.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uri-template==1.3.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.34.0
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==16.0
xgboost==3.2.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
